requirements.txt
streamlit>=1.32.0
xai-sdk>=0.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
plotly>=5.20.0